build10Dec/requirements.txt
aiohttp==3.8.6
aiohttp-cors==0.7.0
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
aubio==0.4.9
audioread==3.0.1
awesomeversion==23.11.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.5
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.13.1
flux-led==1.0.4
fonttools==4.43.1
frozenlist==1.4.0
future==0.18.3
geographiclib==2.0
geopy==2.4.0
icmplib==3.0.4
idna==3.4
ifaddr==0.2.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jplephem==2.19
jupyter_client==8.6.0
jupyter_core==5.5.0
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.3
ledfx==2.0.74
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
msgpack==1.0.7
multidict==5.0.2
nest-asyncio==1.5.8
numba==0.58.1
numpy==1.26.0
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
prompt-toolkit==3.0.40
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
pygame==2.5.2
Pygments==2.16.1
pyinput==0.3.2
pynput==1.7.6
pyparsing==3.1.1
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtChart==5.15.6
PyQtChart-Qt5==5.15.2
pyserial==3.5
pystray==0.19.5
python-dateutil==2.8.2
python-mbedtls==2.7.1
python-osc==1.8.3
python-rtmidi==1.5.7
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
redis==5.0.1
requests==2.28.2
requests-file==1.5.1
sacn==1.6.4
samplerate==0.1.0
savify==2.3.4
scikit-learn==1.3.2
scipy==1.11.3
screen-brightness-control==0.22.1
seaborn==0.13.0
sentry-sdk==1.14.0
sgp4==2.22
six==1.16.0
skyfield==1.46
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spotipy==2.23.0
stack-data==0.6.3
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
tldextract==5.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
validators==0.22.0
voluptuous==0.12.2
wcwidth==0.2.9
webcolors==1.13
WMI==1.5.1
yarl==1.9.2
zeroconf==0.39.4
